feat: 프로젝트 직선레일 등록 페이지, 컴포넌트 분리 - 기존 : 클라이언트 컴포넌트로 전체 처리 - 변경 :  최상위 서버 컴포넌트에서, 클라이언트 컴포넌트 처리하도록 수정
